Sync quaternion module structure with complex module
