+++ Sources/QuaternionModule/Quaternion+ElementaryFunctions.swift
@@ -1,8 +1,8 @@
-//===--- ElementaryFunctions.swift ----------------------------*- swift -*-===//
+//===--- Quaternion+ElementaryFunctions.swift -----------------*- swift -*-===//
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2019-2021 Apple Inc. and the Swift project authors
+// Copyright (c) 2019 - 2022 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
