+++ Sources/QuaternionModule/Quaternion.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift Numerics open source project
 //
-// Copyright (c) 2019 - 2020 Apple Inc. and the Swift Numerics project authors
+// Copyright (c) 2019 - 2022 Apple Inc. and the Swift Numerics project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -87,34 +87,6 @@ extension Quaternion {
     }
   }
 
-  /// The additive identity, with real and *all* imaginary parts zero.
-  ///
-  /// See also:
-  /// -
-  /// - .one
-  /// - .i
-  /// - .j
-  /// - .k
-  /// - .infinity
-  @_transparent
-  public static var zero: Quaternion {
-    Quaternion(from: SIMD4(repeating: 0))
-  }
-
-  /// The multiplicative identity, with real part one and *all* imaginary parts zero.
-  ///
-  /// See also:
-  /// -
-  /// - .zero
-  /// - .i
-  /// - .j
-  /// - .k
-  /// - .infinity
-  @_transparent
-  public static var one: Quaternion {
-    Quaternion(from: SIMD4(0,0,0,1))
-  }
-
   /// The quaternion with the imaginary unit **i** one, i.e. `0 + i + 0j + 0k`.
   ///
   /// See also:
@@ -169,12 +141,6 @@ extension Quaternion {
   @_transparent
   public static var infinity: Quaternion {
     Quaternion(.infinity)
-  }
-
-  /// The conjugate of this quaternion.
-  @_transparent
-  public var conjugate: Quaternion {
-    Quaternion(from: components * [-1, -1, -1, 1])
   }
 
   /// True if this value is finite.
@@ -372,29 +338,6 @@ extension Quaternion {
   public init(real: RealType, imaginary x: RealType, _ y: RealType, _ z: RealType) {
     self.init(real: real, imaginary: SIMD3(x, y, z))
   }
-
-  /// The quaternion with specified real part and zero imaginary part.
-  ///
-  /// Equivalent to `Quaternion(RealType(real))`.
-  @inlinable
-  public init<Other: BinaryInteger>(_ real: Other) {
-    self.init(RealType(real))
-  }
-
-  /// The quaternion with specified real part and zero imaginary part,
-  /// if it can be constructed without rounding.
-  @inlinable
-  public init?<Other: BinaryInteger>(exactly real: Other) {
-    guard let real = RealType(exactly: real) else { return nil }
-    self.init(real)
-  }
-
-  public typealias IntegerLiteralType = Int
-
-  @inlinable
-  public init(integerLiteral value: Int) {
-    self.init(RealType(value))
-  }
 }
 
 extension Quaternion where RealType: BinaryFloatingPoint {
@@ -419,102 +362,5 @@ extension Quaternion where RealType: BinaryFloatingPoint {
         let r = RealType(exactly: other.components.w)
     else { return nil }
     self.init(from: SIMD4(x, y, z, r))
-  }
-}
-
-// MARK: - Conformance to Hashable and Equatable
-extension Quaternion: Hashable {
-  /// Returns a Boolean value indicating whether two values are equal.
-  ///
-  /// - Important:
-  ///   Quaternions are frequently used to represent 3D transformations. It's
-  ///   important to be aware that, when used this way, any quaternion and its
-  ///   negation represent the same transformation, but they do not compare
-  ///   equal using `==` because they are not the same quaternion. You can
-  ///   compare quaternions as 3D transformations using `equals(as3DTransform:)`.
-  @_transparent
-  public static func == (lhs: Quaternion, rhs: Quaternion) -> Bool {
-    // Identify all numbers with either component non-finite as a single "point at infinity".
-    guard lhs.isFinite || rhs.isFinite else { return true }
-    // For finite numbers, equality is defined componentwise. Cases where
-    // only one of lhs or rhs is infinite fall through to here as well, but this
-    // expression correctly returns false for them so we don't need to handle
-    // them explicitly.
-    return lhs.components == rhs.components
-  }
-
-  /// Returns a Boolean value indicating whether the 3D transformation of the
-  /// two quaternions are equal.
-  ///
-  /// Use this method to test for equality of the 3D transformation properties
-  /// of quaternions; where for any quaternion `q`, its negation represent the
-  /// same 3D transformation; i.e. `q.equals(as3DTransform: q)` as well as
-  /// `q.equals(as3DTransform: -q)` are both `true`.
-  ///
-  /// - Parameter other: The value to compare.
-  /// - Returns: True if the 3D transformation of this quaternion equals `other`.
-  @_transparent
-  public func equals(as3DTransform other: Quaternion) -> Bool {
-    // Identify all numbers with either component non-finite as a single "point at infinity".
-    guard isFinite || other.isFinite else { return true }
-    // For finite numbers, equality is defined componentwise. Cases where only
-    // one of lhs or rhs is infinite fall through to here as well, but this
-    // expression correctly returns false for them so we don't need to handle
-    // them explicitly.
-    return components == other.components || components == -other.components
-  }
-
-  @_transparent
-  public func hash(into hasher: inout Hasher) {
-    // There are two equivalence classes to which we owe special attention:
-    // All zeros should hash to the same value, regardless of sign, and all
-    // non-finite numbers should hash to the same value, regardless of
-    // representation. The correct behavior for zero falls out for free from
-    // the hash behavior of floating-point, but we need to use a
-    // representative member for any non-finite values.
-    // For any normal values we use the "canonical transform" representation,
-    // where real is always non-negative. This allows people who are using
-    // quaternions as rotations to get the expected semantics out of collections
-    // (while unfortunately producing some collisions for people who are not,
-    // but not in too catastrophic of a fashion).
-    if isFinite {
-      canonicalizedTransform.components.hash(into: &hasher)
-    } else {
-      hasher.combine(RealType.infinity)
-    }
-  }
-}
-
-// MARK: - Conformance to Codable
-// FloatingPoint does not refine Codable, so this is a conditional conformance.
-extension Quaternion: Decodable where RealType: Decodable {
-  public init(from decoder: Decoder) throws {
-    try self.init(from: SIMD4(from: decoder))
-  }
-}
-
-extension Quaternion: Encodable where RealType: Encodable {
-  public func encode(to encoder: Encoder) throws {
-    try components.encode(to: encoder)
-  }
-}
-
-// MARK: - Formatting
-extension Quaternion: CustomStringConvertible {
-  public var description: String {
-    guard isFinite else {
-        return "inf"
-    }
-    return "(\(components.w), \(components.x), \(components.y), \(components.z))"
-  }
-}
-
-extension Quaternion: CustomDebugStringConvertible {
-  public var debugDescription: String {
-    let x = String(reflecting: components.x)
-    let y = String(reflecting: components.y)
-    let z = String(reflecting: components.z)
-    let r = String(reflecting: components.w)
-    return "Quaternion<\(RealType.self)>(\(r), \(x), \(y), \(z))"
   }
 }
