+++ Sources/QuaternionModule/Transformation.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift Numerics open source project
 //
-// Copyright (c) 2020 Apple Inc. and the Swift Numerics project authors
+// Copyright (c) 2020 - 2022 Apple Inc. and the Swift Numerics project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -121,35 +121,6 @@ extension Quaternion {
   @_transparent
   public var rotationVector: SIMD3<RealType> {
     axis * angle
-  }
-
-  /// The [polar decomposition][wiki].
-  ///
-  /// Returns the length of this quaternion, phase in radians of range *[0, π]*
-  /// and the rotation axis as SIMD3 vector of unit length.
-  ///
-  /// Edge cases:
-  /// -
-  /// - If the quaternion is zero, length is `.zero` and angle and axis
-  /// are `nan`.
-  /// - If the quaternion is non-finite, length is `.infinity` and angle and
-  /// axis are `nan`.
-  /// - For any length other than `.zero` or `.infinity`, if angle is zero, axis
-  /// is `nan`.
-  ///
-  /// See also:
-  /// -
-  /// - `.angle`
-  /// - `.axis`
-  /// - `.angleAxis`
-  /// - `.rotationVector`
-  /// - `init(length:angle:axis:)`
-  /// - `init(length:phase:axis)`
-  /// - `init(rotation:)`
-  ///
-  /// [wiki]: https://en.wikipedia.org/wiki/Polar_decomposition#Quaternion_polar_decomposition
-  public var polar: (length: RealType, phase: RealType, axis: SIMD3<RealType>) {
-    (length, halfAngle, axis)
   }
 
   /// Creates a unit quaternion specified with [Angle-Axis][wiki] values.
@@ -278,68 +249,6 @@ extension Quaternion {
     }
   }
 
-  /// Creates a quaternion specified with [polar coordinates][wiki].
-  ///
-  /// This initializer reads given `length`, `phase` and `axis` values and
-  /// creates a quaternion of equal rotation properties and specified *length*
-  /// using the following equation:
-  ///
-  ///     Q = (cos(phase), axis * sin(phase)) * length
-  ///
-  /// - Note: `axis` must be of unit length, or an assertion failure occurs.
-  ///
-  /// Edge cases:
-  /// -
-  /// - Negative lengths are interpreted as reflecting the point through the origin, i.e.:
-  ///   ```
-  ///   Quaternion(length: -r, phase: θ, axis: axis) == -Quaternion(length: r, phase: θ, axis: axis)
-  ///   ```
-  /// - For any `θ` and any `axis`, even `.infinity` or `.nan`:
-  ///   ```
-  ///   Quaternion(length: .zero, phase: θ, axis: axis) == .zero
-  ///   ```
-  /// - For any `θ` and any `axis`, even `.infinity` or `.nan`:
-  ///   ```
-  ///   Quaternion(length: .infinity, phase: θ, axis: axis) == .infinity
-  ///   ```
-  /// - Otherwise, `θ` must be finite, or a precondition failure occurs.
-  ///
-  /// See also:
-  /// -
-  /// - `.angle`
-  /// - `.axis`
-  /// - `.angleAxis`
-  /// - `.rotationVector`
-  /// - `.polar`
-  /// - `init(length:angle:axis:)`
-  /// - `init(rotation:)`
-  ///
-  /// [wiki]: https://en.wikipedia.org/wiki/Polar_decomposition#Quaternion_polar_decomposition
-  @inlinable
-  public init(length: RealType, phase: RealType, axis: SIMD3<RealType>) {
-    guard !length.isZero, length.isFinite else {
-      self = Quaternion(length)
-      return
-    }
-
-    // Length is finite and non-zero, therefore
-    // 1. `phase` must be finite or a precondition failure needs to occur; as
-    //    this is not representable.
-    // 2. `axis` must be of unit length or an assertion failure occurs; while
-    //    "wrong" by definition, it is representable.
-    precondition(
-      phase.isFinite,
-      "Either phase must be finite, or length must be zero or infinite."
-    )
-    assert(
-      // TODO: Replace with `approximateEquality()`
-      abs(.sqrt(axis.lengthSquared)-1) < max(.sqrt(axis.lengthSquared), 1)*RealType.ulpOfOne.squareRoot(),
-      "Given axis must be of unit length."
-    )
-
-    self = Quaternion(halfAngle: phase, unitAxis: axis).multiplied(by: length)
-  }
-
   /// Transforms a vector by this quaternion.
   ///
   /// Quaternions are frequently used to represent three-dimensional
@@ -404,46 +313,5 @@ extension Quaternion {
     }
     let scale = max(abs(vector.max()), abs(vector.min()))
     return act(on: vector/scale) * scale
-  }
-}
-
-// MARK: - Operations for working with polar form
-
-extension Quaternion {
-  /// The half rotation angle in radians within *[0, π]* range.
-  ///
-  /// Edge cases:
-  /// -
-  /// If the quaternion is zero or non-finite, halfAngle is `nan`.
-  @usableFromInline @inline(__always)
-  internal var halfAngle: RealType {
-    guard isFinite else { return .nan }
-    guard imaginary != .zero else {
-      // A zero quaternion does not encode transformation properties.
-      // If imaginary is zero, real must be non-zero or nan is returned.
-      return real.isZero ? .nan : .zero
-    }
-
-    // If lengthSquared computes without over/underflow, everything is fine
-    // and the result is correct. If not, we have to do the computation
-    // carefully and unscale the quaternion first.
-    let lenSq = imaginary.lengthSquared
-    guard lenSq.isNormal else { return divided(by: magnitude).halfAngle }
-    return .atan2(y: .sqrt(lenSq), x: real)
-  }
-
-  /// Creates a new quaternion from given half rotation angle about given
-  /// rotation axis.
-  ///
-  /// The angle-axis values are transformed using the following equation:
-  ///
-  ///     Q = (cos(halfAngle), unitAxis * sin(halfAngle))
-  ///
-  /// - Parameters:
-  ///   - halfAngle: The half rotation angle
-  ///   - unitAxis: The rotation axis of unit length
-  @usableFromInline @inline(__always)
-  internal init(halfAngle: RealType, unitAxis: SIMD3<RealType>) {
-    self.init(real: .cos(halfAngle), imaginary: unitAxis * .sin(halfAngle))
   }
 }
